Added player movement
ready for animations to be added, should already work. a small note: vertical animation is what takes priority, and flipping animation direction when moving left applies to vertical animations as well. a simple fix I can think of is for horizonal movement to have priority, but this is just a preference thing we can talk about.