Fixed forage scene bug
Added if condition to travel portal in forage scene so that you cant walk out of level before collecting the fragment. Bc I found that its easy to accidentally walk out and if you try to walk back in it causes a bug with the code if you picked up the yellow rock before you walked out.